Changes to get ESP build working with Arduino toolchain.
Almosst there. Build error at the very end

+++ esp/core/spiffs_config.h
@@ -22,8 +22,8 @@
 #include <osapi.h>
 #include "user_config.h"
 #include "flashmem.h"
-typedef signed int s32_t;
-typedef unsigned int u32_t;
+typedef signed long s32_t;
+typedef unsigned long u32_t;
 typedef signed short s16_t;
 typedef unsigned short u16_t;
 typedef signed char s8_t;
